+++ temp/db.seeds.ts
@@ -1,7 +1,7 @@
 import { Database } from '@database';
 import { IUser, UserRole } from '@models/user.model';
-import { INode } from '@models/node.model';
-import { v4 as uuidv4 } from 'uuid';
+import { get as config } from 'config';
+import { hashSync, genSaltSync } from 'bcrypt';
 
 export class DatabaseSeed {
 
@@ -13,13 +13,6 @@ export class DatabaseSeed {
       console.log('Creating database seed');
       const user = await this.seedUser();
       console.log('Created user:', user);
-      const node = await this.seedNode(user.id, 'Dev Node Alfa', 'AA:AA:AA:AA:AA:AA');
-    }
-    if (!await Database.Models.Node.exists({ macAddress: 'B4:D5:BD:9B:5B:4F' })) {
-      Database.Models.User.findOne({ email: 'admin@farmlab.team' }).exec(async (err, user) => {
-        const tomsNode = await this.seedNode(user._id, 'Tom\'s Node', 'B4:D5:BD:9B:5B:4F');
-        console.log('Created node:', tomsNode);
-      })
     }
   }
 
@@ -28,48 +21,12 @@ export class DatabaseSeed {
       const user = new Database.Models.User({
         firstname: 'FarmLab',
         lastname: 'Administrator',
-        email: 'admin@farmlab.team',
-        password: '$2b$12$y7L14XYnseL2F6AUL3AE3.c8l5SXwT5zEKSFYj.hwJMD2MWSYYi/i',
+        email: config('admin.email'),
+        password: hashSync(config('admin.password'), genSaltSync()),
         role: UserRole.admin
       });
       user.save().then(user => { resolve(user) }).catch(error => { reject(error) });
     });
-  }
-
-  async seedNode(userid: string, label: string, mac: string) {
-    const sensorsTypes = ['airtemp', 'watertemp', 'lightstr', 'airhumidity', 'waterph'];
-    const actuatorTypes = ['lightint', 'flowpump', 'foodpump'];
-    return new Promise<INode>(async (resolve, reject) => {
-      let sensorPromises = [];
-      sensorsTypes.forEach(sensor => {
-        let sensorShema = new Database.Models.Sensor({
-          type: sensor
-        });
-        sensorPromises.push(sensorShema.save());
-      });
-
-      let actuatorPromises = [];
-      actuatorTypes.forEach(actuator => {
-        let actuatorShema = new Database.Models.Actuator({
-          type: actuator
-        });
-        actuatorPromises.push(actuatorShema.save());
-      });
-
-      let sensorObjects = await Promise.all(sensorPromises);
-      let actuatorObjects = await Promise.all(actuatorPromises);
-
-      const node = new Database.Models.Node({
-        label:  label,
-        macAddress: mac,
-        authorizationKey: uuidv4(),
-        members: userid,
-        sensors: sensorObjects,
-        actuators: actuatorObjects,
-      });
-      node.save().then(user => { resolve(user) }).catch(error => { reject(error) });
-
-    })
   }
